pyproject.toml
[tool.poetry]
name = "aldangold-page-loader"
version = "0.1.2"
description = "The console page loader "
authors = ["aldangold <meloshnikoff@gmail.com>"]
packages = [
  { include = "page_loader" },
]
repository = "https://github.com/aldangold/python-project-lvl3"
classifiers = [
  "Environment :: Console"
]


[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.24.0"


[tool.poetry.dev-dependencies]
pytest-cov = "^2.10.0"
flake8 = "^3.8.3"


[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"


[[tool.poetry.source]]
name = "ppt"
url = "https://test.pypi.org/legacy/"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
